resourceLocationを移動
